+++ client/owl-app/package.json
@@ -17,12 +17,14 @@
     "axios": "^1.3.6",
     "bootstrap": "^5.2.3",
     "buffer": "^6.0.3",
+    "fontawesome": "^5.6.3",
     "html-react-parser": "^3.0.15",
     "http-proxy-middleware": "^2.0.6",
     "node-html-parser": "^6.1.5",
     "react": "^18.2.0",
     "react-bootstrap": "^2.7.4",
     "react-dom": "^18.2.0",
+    "react-icons": "^4.8.0",
     "react-router-dom": "^6.10.0",
     "react-scripts": "5.0.1",
     "styled-components": "^5.3.9",
@@ -57,5 +59,3 @@
     "snowpack": "^3.8.8"
   }
 }
-
-
